Add RoBERTa encoding preprocessing script
This replaces what what previously llm.preprocess.encode.py
